Add/Mod(Lantern & Player) : Created lantern class and configured lantern sockets for player
Duplicated an engine mesh as a testing player mesh. Created lantern sockets in that mesh, 1:held, 2:stowed, 3:rekindle. Also created a lantern Actor class and blueprint. this class finds the player and attaches itself to a default socket.

Sockets are name based which means the names of the sockets will have to be assigned in the Lantern BP. In scripts this will be managed by a Map and an enum of the 3 sockets.