Mod(LanternFlameDepletion & Rekindle) : Lantern now depletes and can rekindle
Point light intensity now augments based on a rate. This rate is changed based on the lantern state. light intensity is augmented by lerping from its current state to a destination value, which is based on an enum, EFireIntensityTeir.

There are 5 values in EFireIntensityTeir not including a zero'd value "snuffed", the light intensity destination is the percentage of one of those 5 values. For now default destination is snuffed and is changed to tier 5 on ELanternState::ELS_RekindleReady.