Mod/Add(Lantern Rekindling & Player) :  Player now requires m1&m2 to rekindle with scroll wheel
Created another input action for rekindling. This action is scroll wheel and increases flame only when lantern is placed, which is determined by a bool. Added a control binding to IA_Lantern_ReadyLantern in IMC_KBMDefault, now can be inputted with m1 & m2 in accordance with the use and toggle stowed inputs.