Mod(Lantern) : Implemented pos lerp for lantern states, but needs refactoring
I set up a timeline in lantern bp, with functionality to lerp the actor to the active socket on a function called when changing lantern states. Once lerp is complete the actor is attached to said socket. This should be refactored because the attachment happens in the bp not code where it should.