Mod/Fix (Lantern & LHCharacter) : Fixed lantern attachment & improved positional lerp
Removed unnecessary mesh1P variable as get mesh is already a function(this was the solution to solve attachment bug). Lantern now has an initialize function which sets the meshwithlanternsockets. Also added a few bp available functions to Lantern to get the mesh with sockets and attach the mesh, so the lerp is handled in the bp and attachment is handled in lantern.cpp.