Add/Mod(InteractionComponent,Player & Lantern) : Player can now Pick up lantern
Created an interaction component that broadcasts a delegate for objects to subscribe to. Lantern has this component in in bp prints text when event is fired, in the lantern cpp when the event is fires the interaction component is destroyed along with the collision shape.

Player has a function to sphere cast which returns the first hit results, and for now in the OnInteractAction() this calls the interact component delegate if the returning result is an actor with the interact component. In this section, if the player does not have a lantern, and they sphere cast the lantern, the lantern is then assigned and attatched.