Mod(Lantern & LHCharacter) : Panic now only increases when lantern is off
Renamed ELS_Held to ELS_InUse and created another lantern state called ELS_Rekindling. Made this state always switch from stowed and switches to ELS_InUse after a delay. Changed the InitializeLantern function to include a panic manager as a paramater and used that panic manager to set Panicking to true only when the Lantern state is InUse.