Mod( LHCaracter, Lantern & PanicManager) : created a debug key, and updated controls
Debug key is 'k' for now. Created  input action for a debug key and added it to the IMC_KBMDefault. Also added a helper function to Panic manager to get panic. and created a debug variable for the character and the lantern so now only on screen messages are shown if the debug mode is on