Mod(PanicManager) : Breath Option Occurs After a Delay after being seen
Altered the LerpPanicMeter to include logic that judges weather the player has been seen. created functions like StartOutOfLineOfSightDelay which starts a delay timer if the player had been seen. And UpdateMassPanicReductionFlag which updates bIsMassPanicReductionEnabled, a bool that determines whether or not the player is able to breathe or reduce panic by a mass value.